Add server info bar entry and plugin IPC
DTR bar indicates current search type and displays search pattern in tooltip, when one is present. When no search is active, tooltip displays instructions to set one.

IPC providers are available to get the search pattern for each type, as well as a unified `TYPE:PATTERN` format (or, in the case of no active search, `TYPE` alone), as well as checking for the presence of any search pattern, clearing the active search, and setting the different types.

Note that using IPC to set a search with an empty string will NOT clear the search, as this is assumed to be an error by the caller. Callers are responsible for determining that they wish to clear the search pattern and using the appropriate IPC call.