Stop drawing overlay in certain conditions
If the player is in a cutscene, between areas, or in character creation, the ESP overlay will be automatically disabled until they are no longer in those conditions. This should prevent annoying overlays from appearing during cutscenes and breaking immersion, and may avoid potential problems with scanning the object table while it's being loaded between zones.